+++ learning/library/butterknife/src/debug/AndroidManifest.xml
