+++ learning/library/realm/src/debug/java/com/github/ttanaka330/learning/todo/DebugApp.kt
@@ -1,0 +1,38 @@
+package com.github.ttanaka330.learning.todo
+
+import android.os.StrictMode
+import com.facebook.stetho.Stetho
+import com.uphyca.stetho_realm.RealmInspectorModulesProvider
+
+class DebugApp : App() {
+
+    override fun onCreate() {
+        setupStrict()
+        super.onCreate()
+        setupStetho()
+    }
+
+    private fun setupStetho() {
+        Stetho.initialize(
+            Stetho.newInitializerBuilder(this)
+                .enableDumpapp(Stetho.defaultDumperPluginsProvider(this))
+                .enableWebKitInspector(RealmInspectorModulesProvider.builder(this).build())
+                .build()
+        )
+    }
+
+    private fun setupStrict() {
+        StrictMode.setThreadPolicy(
+            StrictMode.ThreadPolicy.Builder()
+                .detectAll()
+                .penaltyLog()
+                .build()
+        )
+        StrictMode.setVmPolicy(
+            StrictMode.VmPolicy.Builder()
+                .detectAll()
+                .penaltyLog()
+                .build()
+        )
+    }
+}
