+++ learning/library/realm/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.github.ttanaka330.learning.todo.realm">
+    package="com.github.ttanaka330.learning.todo">
 
     <application
         android:name=".App"
