+++ learning/library/realm/src/main/java/com/github/ttanaka330/learning/todo/App.kt
@@ -1,4 +1,4 @@
-package com.github.ttanaka330.learning.todo.realm
+package com.github.ttanaka330.learning.todo
 
 import android.app.Application
 import io.realm.Realm
