+++ learning/library/realm/src/main/java/com/github/ttanaka330/learning/todo/TaskCompletedFragment.kt
@@ -1,4 +1,4 @@
-package com.github.ttanaka330.learning.todo.realm
+package com.github.ttanaka330.learning.todo
 
 import android.content.DialogInterface
 import android.content.Intent
@@ -12,10 +12,10 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import com.github.ttanaka330.learning.todo.realm.data.Task
-import com.github.ttanaka330.learning.todo.realm.data.TaskRepository
-import com.github.ttanaka330.learning.todo.realm.data.TaskRepositoryDataSource
-import com.github.ttanaka330.learning.todo.realm.widget.ConfirmMessageDialog
+import com.github.ttanaka330.learning.todo.data.Task
+import com.github.ttanaka330.learning.todo.data.TaskRepository
+import com.github.ttanaka330.learning.todo.data.TaskRepositoryDataSource
+import com.github.ttanaka330.learning.todo.widget.ConfirmMessageDialog
 import kotlinx.android.synthetic.main.fragment_task_list.view.*
 
 class TaskCompletedFragment : BaseFragment(), TaskListAdapter.ActionListener {
