+++ learning/library/realm/src/main/java/com/github/ttanaka330/learning/todo/data/TaskObject.kt
@@ -1,4 +1,4 @@
-package com.github.ttanaka330.learning.todo.realm.data
+package com.github.ttanaka330.learning.todo.data
 
 import io.realm.RealmObject
 import io.realm.annotations.PrimaryKey
